create legacy package with new hyper serup
