+++ autogalaxy/legacy/__init__.py
@@ -1,0 +1,1 @@
+from setup import SetupHyper
